+++ src/hooks/index.ts
@@ -1,5 +1,3 @@
 export * from './useAutoCallback';
-export * from './useSearchOptions';
 export * from './useLectureData';
-export * from './useFilteredLectures';
-export * from './usePagination';
+export * from './useSearchWithPagination';
